Add CO2 and tvoc sensors reporting.
I needed to delete and add again the device to HA for it to see new sensors.

NOTE: https://github.com/espressif/esp-matter/issues/911 needs to be fixed for the code to compile.
NOTE: Home assistant matter integration lacks bindings for tVOC so the value will not be visible in the UI.